refacto!: EmojiManager instantiation
Breaking change: EmojiManager is now asking for context + IDeserializer to instiate
