+++ app/src/main/java/io/wax911/emojifysample/App.kt
@@ -15,9 +15,18 @@ class App : Application() {
      * dependency injector framework like as a singleton in `Hilt`, `Dagger` or `Koin`
      */
     internal val emojiManager: EmojiManager by lazy {
-        // should already be initialized if we haven't disabled initialization in manifest
-        // see: https://developer.android.com/topic/libraries/app-startup#disable-individual
-        AppInitializer.getInstance(this)
-            .initializeComponent(EmojiInitializer::class.java)
+        EmojiManager.create(this, CustomEmojiInitializer.MoshiDeserializer())
     }
+
+//    /**
+//     * Application scope bound emojiManager, you could keep a reference to this object in a
+//     * dependency injector framework like as a singleton in `Hilt`, `Dagger` or `Koin`
+//     */
+//    internal val startupEmojiManager: EmojiManager by lazy {
+//        EmojiManager.create(this, CustomEmojiInitializer.MoshiDeserializer())
+//        // should already be initialized if we haven't disabled initialization in manifest
+//        // see: https://developer.android.com/topic/libraries/app-startup#disable-individual
+//        AppInitializer.getInstance(this)
+//            .initializeComponent(EmojiInitializer::class.java)
+//    }
 }
